+++ assembly/constants/addresses.ts
@@ -1,0 +1,8 @@
+import { Address } from "@graphprotocol/graph-ts";
+
+export const WETH = Address.fromString(
+  "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
+);
+export const USDC = Address.fromString(
+  "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
+);
