+++ assembly/constants/units.ts
@@ -1,0 +1,3 @@
+import { integers } from "..";
+
+export const USDC = integers.TEN.pow(6);
